move table functions to separate module
